# api/openapi.yaml
openapi: 3.0.3
info:
  title: Golang Microservice Template API
  description: |
    Production-ready Golang microservice template with hexagonal architecture.
    
    Features:
    - RESTful API for item management
    - JWT authentication support
    - Prometheus metrics
    - OpenTelemetry tracing
    - Kafka event streaming
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Development server
  - url: https://api.example.com
    description: Production server

tags:
  - name: health
    description: Health check endpoints
  - name: items
    description: Item management endpoints

paths:
  /health:
    get:
      tags:
        - health
      summary: Health check
      description: Returns the health status of the service
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: healthy

  /v1/items:
    get:
      tags:
        - items
      summary: List items
      description: Retrieve a paginated list of items
      operationId: listItems
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of items to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: offset
          in: query
          description: Number of items to skip
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Item'
                  limit:
                    type: integer
                  offset:
                    type: integer
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      tags:
        - items
      summary: Create item
      description: Create a new item
      operationId: createItem
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

  /v1/items/{id}:
    get:
      tags:
        - items
      summary: Get item
      description: Retrieve a specific item by ID
      operationId: getItem
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Item ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - items
      summary: Update item
      description: Update an existing item
      operationId: updateItem
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Item ID
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemRequest'
      responses:
        '200':
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

    delete:
      tags:
        - items
      summary: Delete item
      description: Delete an item
      operationId: deleteItem
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Item ID
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Item deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication

  schemas:
    Item:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "123e4567-e89b-12d3-a456-426614174000"
        name:
          type: string
          example: "Example Item"
          minLength: 1
          maxLength: 100
        description:
          type: string
          example: "This is an example item"
        price:
          type: number
          format: double
          minimum: 0
          example: 19.99
        created_at:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2023-01-01T12:00:00Z"
      required:
        - id
        - name
        - price
        - created_at
        - updated_at

    CreateItemRequest:
      type: object
      properties:
        name:
          type: string
          example: "Example Item"
          minLength: 1
          maxLength: 100
        description:
          type: string
          example: "This is an example item"
        price:
          type: number
          format: double
          minimum: 0
          example: 19.99
      required:
        - name
        - price

    UpdateItemRequest:
      type: object
      properties:
        name:
          type: string
          example: "Updated Item"
          minLength: 1
          maxLength: 100
        description:
          type: string
          example: "This is an updated item"
        price:
          type: number
          format: double
          minimum: 0
          example: 29.99
      required:
        - name
        - price

    Error:
      type: object
      properties:
        error:
          type: string
          example: "Error message"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
                example: "name"
              message:
                type: string
                example: "is required"

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation failed"
            errors:
              - field: "name"
                message: "is required"

    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Missing authorization header"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Item not found"

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"